requirements.txt
streamlit
python-docx
Pillow
jinja2
